refactor(extension/podman): extracting WSLVersionCheck
